This commit implements the following features:
* Add basic support for `maxWidth` for Panels. This can be specified just like "minWidth", and defaults to 0
* Add a second argument to the `onUpdate` callback, allowing more seamless integration with external state libraries

Additionally, code has been refactored to break apart Divider, Panel, and Panel Group into standalone components.

Remaining TODO:

* Add tests covering the `maxWidth` and `isComplete` behaviors.
